Modularized source, splited fuzzymatrix and perceptron into different modules Added accuracy statistics evolution during training
