+++ src/fuzzymatrix.py
@@ -1,0 +1,28 @@
+class FuzzyMatrix:
+    def __init__(self):
+        self.matrix = {}
+
+    def add(self, predicted, expected):
+        key = (predicted, expected)
+        self.matrix[key] = self.matrix.get(key, 0) + 1
+
+    def trace(self):
+        matrixList = (self.matrix.items())
+        return sum(map(lambda x: x[1], filter(lambda x: x[0][0] == x[0][1], matrixList)))
+
+    def sum(self):
+        return sum(self.matrix.values())
+
+    def accuracy(self):
+        return self.trace()/float(self.sum())
+
+    def precision(self, label):
+        predictedLabel = sum(map(lambda x: x[1], filter(lambda x: x[0][0] == label, self.matrix.items())))
+        predictedCorrectly = self.matrix[(label, label)]
+        return predictedCorrectly / float(predictedLabel)
+
+    def sensitivity(self, label):
+        expectedLabel = sum(map(lambda x: x[1], filter(lambda x: x[0][1] == label, self.matrix.items())))
+        predictedCorrectly = self.matrix[(label, label)]
+        return predictedCorrectly / float(expectedLabel)
+
